+++ fastapi_errors_plus/__init__.py
@@ -2,6 +2,7 @@
 
 from fastapi_errors_plus.base import BaseErrorDTO, StandardErrorDTO
 from fastapi_errors_plus.error_doc import ErrorDoc
+from fastapi_errors_plus.error_profile import ErrorProfile
 from fastapi_errors_plus.errors import Errors
 from fastapi_errors_plus.protocol import ErrorDTO, ErrorDTOLike, LegacyErrorDTO
 
@@ -11,7 +12,8 @@ __all__ = [
     "LegacyErrorDTO",
     "ErrorDTOLike",
     "ErrorDoc",
+    "ErrorProfile",
     "BaseErrorDTO",
     "StandardErrorDTO",
 ]
-__version__ = "0.8.0"
+__version__ = "0.9.0"